file upload

+++ routes/dishRouter.js
@@ -1,15 +1,18 @@
 const express = require('express');
 const bodyParser = require('body-parser');
-const mongoose = require('mongoose');
+
+const authenticate = require('../authenticate')
 
 const Dishes = require('../models/dishes');
 const dishRouter = express.Router();
 
 dishRouter.use(bodyParser.json());
+
 dishRouter.route('/')
 .get((req,res,next) => {
 
         Dishes.find({})
+        .populate('comments.author')
         .then((dishes) => {
             res.statusCode = 200;
             res.setHeader('Content-Type','application/json');
@@ -17,23 +20,37 @@ dishRouter.route('/')
         },(err)=>next(err))
          .catch((err) => next(err));
 })
-.post((req,res,next) => {
-
+.post(authenticate.verifyUser ,(req,res,next) => {
+    if(authenticate.verifyAdmin(req.user)==false){
+        
+        err = new Error('Only Admins are allowed to access');
+        err.status = 404;
+        return next(err);
+    }
+    if(authenticate.verifyAdmin(req.user)==true){
     Dishes.create(req.body)
     .then((dish)=>{
+
         console.log('Dish Created',dish);
         res.statusCode = 200;
         res.setHeader('Content-Type','application/json');
         res.json(dish);
     },(err)=>next(err))
     .catch((err) => next(err));
-})
-.put((req,res,next) => {
+}})
+.put(authenticate.verifyUser,(req,res,next) => {
 
-    res.statusCode= 404;
+    res.statusCode= 403;
     res.end('PUT operation not supported');
 })
-.delete((req,res,next) => {
+.delete(authenticate.verifyUser,(req,res,next) => {
+    if(authenticate.verifyAdmin(req.user)==false){
+        
+        err = new Error('Only Admins are allowed to access');
+        err.status = 404;
+        return next(err);
+    }
+    if(authenticate.verifyAdmin(req.user)==true){
     Dishes.remove({})
     .then((resp)=>{
         res.statusCode = 200;
@@ -42,7 +59,7 @@ dishRouter.route('/')
         
     },(err)=>next(err))
     .catch((err) => next(err));
-});
+}});
 
 
 
@@ -50,6 +67,7 @@ dishRouter.route('/:dishId')
 .get((req,res,next) => {
     
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         res.statusCode = 200;
         res.setHeader('Content-Type','application/json');
@@ -58,13 +76,19 @@ dishRouter.route('/:dishId')
      .catch((err) => next(err));
 
 })
-.post((req,res,next) => {
+.post(authenticate.verifyUser,(req,res,next) => {
 
     res.statusCode= 404;
     res.end('POST operation not supported on /dishes/'+req.params.dishId);
 })
-.put((req,res,next) => {
-
+.put(authenticate.verifyUser,(req,res,next) => {
+    if(authenticate.verifyAdmin(req.user)==false){
+        
+        err = new Error('Only Admins are allowed to access');
+        err.status = 404;
+        return next(err);
+    }
+    if(authenticate.verifyAdmin(req.user)==true){
     Dishes.findByIdAndUpdate(req.params.dishId,{
         $set:req.body
     },{new:true})
@@ -75,8 +99,15 @@ dishRouter.route('/:dishId')
     },(err)=>next(err))
      .catch((err) => next(err));
 
-})
-.delete((req,res,next) => {
+}})
+.delete(authenticate.verifyUser,(req,res,next) => {
+    if(authenticate.verifyAdmin(req.user)==false){
+        
+        err = new Error('Only Admins are allowed to access');
+        err.status = 404;
+        return next(err);
+    }
+    if(authenticate.verifyAdmin(req.user)==true){
     Dishes.findByIdAndRemove(req.params.dishId)
     .then((resp)=>{
         res.statusCode = 200;
@@ -86,7 +117,7 @@ dishRouter.route('/:dishId')
     },(err)=>next(err))
     .catch((err) => next(err));
 
-});
+}});
 
 
 
@@ -95,6 +126,7 @@ dishRouter.route('/:dishId/comments')
 .get((req,res,next) => {
 
         Dishes.findById(req.params.dishId)
+        .populate('comments.author')
         .then((dish) => {
             if(dish!=null){
                 res.statusCode = 200;
@@ -109,18 +141,23 @@ dishRouter.route('/:dishId/comments')
         },(err)=>next(err))
          .catch((err) => next(err));
 })
-.post((req,res,next) => {
+.post(authenticate.verifyUser, (req,res,next) => {
 
     Dishes.findById(req.params.dishId)
     .then((dish)=>{
         if(dish!=null){
-           
+           req.body.author =req.user._id; 
             dish.comments.push(req.body);
             dish.save()
             .then((dish)=>{
-                res.statusCode = 200;
-                res.setHeader('Content-Type','application/json');
-                res.json(dish.comments);
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then((dish)=>{
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type','application/json');
+                    res.json(dish.comments); 
+                })
+
             },(err)=>next(err));
         }
         else{
@@ -131,12 +168,19 @@ dishRouter.route('/:dishId/comments')
     },(err)=>next(err))
     .catch((err) => next(err));
 })
-.put((req,res,next) => {
+.put(authenticate.verifyUser,(req,res,next) => {
 
     res.statusCode= 403;
     res.end('PUT operation not supported om /dish/'+req.params.dishId+'/comments');
 })
-.delete((req,res,next) => {
+.delete(authenticate.verifyUser,(req,res,next) => {
+    if(authenticate.verifyAdmin(req.user)==false){
+        
+        err = new Error('Only Admins are allowed to access');
+        err.status = 404;
+        return next(err);
+    }
+    if(authenticate.verifyAdmin(req.user)==true){
     Dishes.findById(req.params.dishId)
     .then((dish)=>{
         if(dish!=null){
@@ -159,7 +203,7 @@ dishRouter.route('/:dishId/comments')
         }
     },(err)=>next(err))
     .catch((err) => next(err));
-});
+}});
 
 
 
@@ -167,6 +211,7 @@ dishRouter.route('/:dishId/comments/:commentId')
 .get((req,res,next) => {
     
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         if(dish!=null&&dish.comments.id(req.params.commentId)!=null){
             res.statusCode = 200;
@@ -187,15 +232,25 @@ dishRouter.route('/:dishId/comments/:commentId')
      .catch((err) => next(err));
 
 })
-.post((req,res,next) => {
+.post(authenticate.verifyUser,(req,res,next) => {
 
     res.statusCode= 404;
     res.end('POST operation not supported on /dishes/'+req.params.dishId+'/comments/'+req.params.commentId);
 })
-.put((req,res,next) => {
+.put(authenticate.verifyUser,(req,res,next) => {
+    
+ console.log(req.user._id);
     Dishes.findById(req.params.dishId)
+    .populate('comments.author')
     .then((dish) => {
         if(dish!=null&&dish.comments.id(req.params.commentId)!=null){
+            if(!(req.user._id.equals(dish.comments.id(req.params.commentId).author._id)))
+            {
+                console.log(req.user._id+" "+dish.comments.id(req.params.commentId).author._id)
+                err = new Error("Not authorised to edit");
+                err.status = 403;
+                return next(err);
+            }
             if(req.body.rating){
                 dish.comments.id(req.params.commentId).rating = req.body.rating;
 
@@ -206,9 +261,14 @@ dishRouter.route('/:dishId/comments/:commentId')
             
             dish.save()
             .then((dish)=>{
-                res.statusCode = 200;
-                res.setHeader('Content-Type','application/json');
-                res.json(dish.comments);
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then((dish)=>{
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type','application/json');
+                    res.json(dish.comments); 
+                })
+
             },(err)=>next(err));
         }
         else if(dish ==null){
@@ -224,17 +284,28 @@ dishRouter.route('/:dishId/comments/:commentId')
     },(err)=>next(err))     .catch((err) => next(err));
 
 })
-.delete((req,res,next) => {
+.delete(authenticate.verifyUser,(req,res,next) => {
     Dishes.findById(req.params.dishId)
     .then((dish)=>{
         if(dish!=null&&dish.comments.id(req.params.commentId)!=null){
+            if(!(req.user._id==(dish.comments.author._id)))
+            {
+                err = new Error("Not authorised to edit");
+                err.status = 403;
+                return next(err);
+            }
                       
                dish.comments.id(req.params.commentId).remove();
            dish.save()
             .then((dish)=>{
-                res.statusCode = 200;
-                res.setHeader('Content-Type','application/json');
-                res.json(dish);
+                Dishes.findById(dish._id)
+                .populate('comments.author')
+                .then((dish)=>{
+                    res.statusCode = 200;
+                    res.setHeader('Content-Type','application/json');
+                    res.json(dish.comments); 
+                })
+
             },(err)=>next(err));
         }
         else if(dish ==null){
